chore: resolve duplicate IssueCard naming collision (maintainers vs shared/ui)
Two entirely unrelated components shared the exact name IssueCard, in
different directories, with no shared code, different prop shapes, and
different rendering logic:
- src/features/maintainers/components/issues/IssueCard.tsx: single
  `issue: Issue` prop, hand-rolled rendering for the maintainers issue-list
  sidebar specifically.
- src/shared/components/ui/IssueCard.tsx: a much more general, prop-driven,
  data-testid-equipped component supporting a `variant="recommended"` mode,
  used across most of the app.

Fully unifying them would mean mapping Issue's fields onto the shared
component's much larger, more generic prop surface and re-verifying visual
parity across every field - a substantially larger, higher-risk refactor
than resolving the naming collision itself required. Took the
lower-risk option the issue explicitly allows instead: renamed the
maintainers-local one to MaintainerIssueCard (and its test file to match),
updated its sole importer (IssueListSidebar.tsx) and that file's test mock,
and left a short cross-referencing comment in both IssueCard files so
future readers understand the split is intentional, not an accidental
collision.

Closes #659
